move the logic of subscribing to the redux store to the DiagramTab component, fix display of negative values
